separate Course resource DTOs under LMS.Shared.DTOs.CourseDtos namespace
To try to minimize merge conflicts when doing concurrent work on the Course resource.
