+++ LMS.Shared/DTOs/CourseDtos/CourseDtos.cs
@@ -1,0 +1,19 @@
+﻿namespace LMS.Shared.DTOs.CourseDtos
+{
+    public class CourseUpsertDto
+    {
+        public string Name { get; set; } = string.Empty;
+        public string Description { get; set; } = string.Empty;
+        public DateTime StartDate { get; set; }
+        public DateTime EndDate { get; set; }
+    }
+
+    public class CourseReadDto
+    {
+        public int Id { get; set; }
+        public string Name { get; set; } = string.Empty;
+        public string Description { get; set; } = string.Empty;
+        public DateTime StartDate { get; set; }
+        public DateTime EndDate { get; set; }
+    }
+}
