+++ LMS.Shared/DTOs/CourseDtos/CourseQueryDtos.cs
@@ -1,0 +1,10 @@
+﻿namespace LMS.Shared.DTOs.CourseDtos;
+
+public class CoursesQueryDto
+{
+}
+
+public class CoursesQueryResultDto
+{
+
+}
