+++ Service.Contracts/ICourseService.cs
@@ -1,4 +1,4 @@
-﻿using LMS.Shared.DTOs;
+﻿using LMS.Shared.DTOs.CourseDtos;
 
 namespace Service.Contracts
 {
